requirements.txt
daphne==2.4.1
dataclasses-json==0.3.7
django-filter==2.2.0
django==3.0.2
djangorestframework==3.11.0
factory_boy==2.12.0
markdown==3.1.1
psycopg2-binary==2.8.4
pytest-django==3.7.0
pytest-mock==2.0.0
python-rapidjson==0.9.1
requests==2.22.0
simpleeval==0.9.10
whitenoise==5.0.1